+++ PgRoutiner/Program/ArgsInclude.cs
@@ -5,7 +5,7 @@ namespace PgRoutiner
 {
     static partial class Program
     {
-        private static bool ArgsInclude(string[] args, params string[] values)
+        public static bool ArgsInclude(string[] args, params string[] values)
         {
             var lower = values.Select(v => v.ToLower()).ToList();
             var upper = values.Select(v => v.ToUpper()).ToList();