+++ CHANGES.md
@@ -1,10 +1,12 @@
-
 # Version history
 
 ## 2.1.0
 
+- Automatic package upgrade with confirmation
+- Overwrite confirmation with `AskOverwrite` option
 
 ## 2.0.0
 
 - Complete source code builder engine rewrite
-- Support for 
+- Support for Norm 3.1+ versions
+- Indentation support with `Ident` option